IM DUN
omg we have a fully working game and made all the dialog work and game
repeats until all dialog is found and then it hard exits so we need to
come up with better ending and maybe put sounds and music and im going
to bed. gg. definitly re.
ps - awesome job dude.